Function relay_system::compat::send

pub async fn send<A, M>(addr: Addr<A>, msg: M) -> Result<M::Result, MailboxError>where
    A: Actor + Send,
    M: Message + Send + 'static,
    M::Result: Send,
    A: Handler<M>,
    A::Context: ToEnvelope<A, M>,

Sends a message to an actor using actix 0.7 from a tokio 1.0 context.

The message is internally forwarded through a tokio 0.1 runtime in order to avoid panics when the channel queues fill up and tasks are getting parked.

Panics

Panics if this is invoked outside of the Controller.