Crate relay_system

Foundational system components for Relay’s services.

Relay’s system is based on [tokio]. To use any of these components, ensure a tokio runtime is available.

Services

The basic building block in Relay are asynchronous Services. Each service implements an Interface, which consists of one or many messages that can be sent to the service using its Addr. See the docs of these types for more information on how to implement and use them.

Shutdown

The static Controller service can listen for process signals and initiate a graceful shutdown. Note that services must check a ShutdownHandle to receive these signals. See the struct level documentation for more information.

Structs

Addr

The address of a Service.

AsyncResponse

Message response resulting in an asynchronous Request.

BroadcastChannel

A channel that broadcasts values to attached senders.

BroadcastRequest

The request when sending an asynchronous message to a service.

BroadcastResponse

Variation of AsyncResponse that efficiently broadcasts responses to many requests.

BroadcastSender

Sends a message response from a service back to the waiting BroadcastRequest.

Controller

Service to start and gracefully stop the system runtime.

NoResponse

Message response for fire-and-forget messages with no output.

Receiver

Inbound channel for messages sent through an Addr.

Recipient

An address to a Service implementing any interface that takes a given message.

Request

The request when sending an asynchronous message to a service.

SendError

An error when sending a message to a service fails.

Sender

Sends a message response from a service back to the waiting Request.

Shutdown

Shutdown request message sent by the Controller to subscribed actors.

ShutdownHandle

Notifies a service about an upcoming shutdown.

Enums

ShutdownMode

Determines how to shut down the Relay system.

Traits

FromMessage

Declares a message as part of an Interface.

Interface

A message interface for services.

MessageResponse

Response behavior of an Interface message.

Service

An asynchronous unit responding to messages.

Functions

channel

Creates an unbounded channel for communicating with a Service.