Crate relay_system

Foundational system components for Relay’s services.

Actors require an actix system to run. The system can be started using the Controller actor, which will also listen for shutdown signals and trigger a graceful shutdown. Note that actors must implement a handler for the Shutdown message and register with the controller to receive this signal. See the struct level documentation for more information.

See the Controller struct for more information.

Modules

compat

Compatibility layer for bridging between actix and tokio 1.0.

Structs

Addr

The address of a Service.

AsyncResponse

Message response resulting in an asynchronous Request.

Configure

Configures the Controller with new parameters.

Controller

Actor to start and gracefully stop an actix system.

NoResponse

Message response for fire-and-forget messages with no output.

Request

The request when sending an asynchronous message to a service.

SendError

An error when sending a message to a service fails.

Sender

Sends a message response from a service back to the waiting Request.

Shutdown

Shutdown request message sent by the Controller to subscribed actors.

Signal

Process signal message

Subscribe

Subscribtion message for Shutdown events.

SubscribeV2

Internal message to handle subscribing to the Shutdown message.

Enums

SignalType

Different types of process signals

Traits

FromMessage

Declares a message as part of an Interface.

Interface

A message interface for services.

MessageResponse

Response behavior of an Interface message.

Service

An asynchronous unit responding to messages.

Functions

channel

Creates an unbounded channel for communicating with a Service.

Type Definitions

Receiver

Inbound channel for messages sent through an Addr.