Struct relay_system::Controller

pub struct Controller { /* private fields */ }

Service to start and gracefully stop an the system runtime.

This service contains a static run method which will run a tokio system and block the current thread until the system shuts down again.

It starts with default configuration. To change this configuration, send the Configure message.

To shut down more gracefully, other actors can register with Controller::shutdown_handle. When a shutdown signal is sent to the process, they will receive a Shutdown message with an optional timeout. They can respond with a future, after which they will be stopped. Once all registered actors have stopped successfully, the entire system will stop.

Example

// TODO: This example is outdated and needs to be updated when the Controller is rewritten.
use actix::prelude::*;
use relay_system::{Controller, Shutdown};

struct MyActor;

impl Actor for MyActor {
    type Context = Context<Self>;

    fn started(&mut self, context: &mut Self::Context) {
        Controller::subscribe(context.address());
    }
}

impl Handler<Shutdown> for MyActor {
    type Result = Result<(), ()>;

    fn handle(&mut self, message: Shutdown, _context: &mut Self::Context) -> Self::Result {
        // Implement custom logic here
        Ok(())
    }
}

Controller::run(|| -> Result<(), ()> {
    MyActor.start();
    Ok(())
}).unwrap();

Implementations

impl Controller

pub fn from_registry() -> Addr<Self>

Get actor’s address from system registry.

Panics

This method panics when it is invoked outside of a controller context (Controller::run).

pub fn run<F, R, E>(factory: F) -> Result<(), E>where
F: FnOnce() -> Result<R, E> + 'static + Sync + Send,

Runs the factory to start actors.

The function accepts the old tokio 0.x [actix::SystemRunner] and the reference to [tokio::runtime::Handle] from the new tokio 1.x runtime, which we enter before the factory is run, to make sure that two systems, old and new one, are available.

The factory may be used to start actors in the actix system before it runs. If the factory returns an error, the actix system is not started and instead an error returned. Otherwise, the system blocks the current thread until a shutdown signal is sent to the server and all actors have completed a graceful shutdown.

pub fn shutdown_handle() -> ShutdownHandle

Returns a handle to receive shutdown notifications.