Struct relay_system::Controller

pub struct Controller { /* private fields */ }

Actor to start and gracefully stop an actix system.

This actor contains a static run method which will run an actix system and block the current thread until the system shuts down again.

This actor starts with default configuration. To change this configuration, send the Configure message.

To shut down more gracefully, other actors can register with the Subscribe message. When a shutdown signal is sent to the process, they will receive a Shutdown message with an optional timeout. They can respond with a future, after which they will be stopped. Once all registered actors have stopped successfully, the entire system will stop.

Example

use actix::prelude::*;
use relay_system::{Controller, Shutdown};

struct MyActor;

impl Actor for MyActor {
    type Context = Context<Self>;

    fn started(&mut self, context: &mut Self::Context) {
        Controller::subscribe(context.address());
    }
}

impl Handler<Shutdown> for MyActor {
    type Result = Result<(), ()>;

    fn handle(&mut self, message: Shutdown, _context: &mut Self::Context) -> Self::Result {
        // Implement custom logic here
        Ok(())
    }
}

Controller::run(|| -> Result<(), ()> {
    MyActor.start();
    Ok(())
}).unwrap();

Implementations

impl Controller

pub fn from_registry() -> Addr<Self>

Get actor’s address from system registry.

pub fn run<F, R, E>(factory: F) -> Result<(), E> where
F: FnOnce() -> Result<R, E>,

Starts an actix system and runs the factory to start actors.

The factory may be used to start actors in the actix system before it runs. If the factory returns an error, the actix system is not started and instead an error returned. Otherwise, the system blocks the current thread until a shutdown signal is sent to the server and all actors have completed a graceful shutdown.

pub fn subscribe<A>(addr: Addr<A>) where
A: Handler<Shutdown>,
A::Context: ToEnvelope<A, Shutdown>,

Subscribes the provided actor to the Shutdown signal of the system controller.

pub async fn subscribe_v2() -> Receiver<Option<Shutdown>>

Subscribes to the Shutdown message to handle graceful shutdown.

Returns a receiver for the Shutdown message, to be used to gracefully shutdown. This sends a message to the Controller actor so will block until this actor is running.

TODO(tobias): The receiver of this message can not yet signal they have completed shutdown.