Trait relay_system::MessageResponse

pub trait MessageResponse {
    type Sender;
    type Output;

    fn channel() -> (Self::Sender, Self::Output);
}

Response behavior of an Interface message.

It defines how a service handles and responds to interface messages, such as through asynchronous responses or fire-and-forget without responding. FromMessage implementations declare this behavior on the interface.

See FromMessage for more information on how to use this trait.

Required Associated Types

type Sender

Sends responses from the service back to the waiting recipient.

type Output

The type returned from Addr::send.

This type can be either synchronous and asynchronous based on the responder.

Required Methods

fn channel() -> (Self::Sender, Self::Output)

Returns the response channel for an interface message.