Trait relay_system::Service

pub trait Service: Sized {
    type Interface: Interface;

    fn spawn_handler(self, rx: Receiver<Self::Interface>);

    fn start(self) -> Addr<Self::Interface> { ... }
}

An asynchronous unit responding to messages.

Services receive messages conforming to some Interface through an Addr and handle them one by one. Internally, services are free to concurrently process these messages or not, most probably should.

Individual messages can have a response which will be sent once the message is handled by the service. The sender can asynchronously await the responses of such messages.

To start a service, create an instance of the service and use Service::start.

Implementing Services

The standard way to implement services is through the run function. It receives an inbound channel for all messages sent through the service’s address. Note that this function is synchronous, so that this needs to spawn at least one task internally:

use relay_system::{FromMessage, Interface, NoResponse, Receiver, Service};

struct MyMessage;

impl Interface for MyMessage {}

impl FromMessage<Self> for MyMessage {
    type Response = NoResponse;

    fn from_message(message: Self, _: ()) -> Self {
        message
    }
}

struct MyService;

impl Service for MyService {
    type Interface = MyMessage;

    fn spawn_handler(self, mut rx: Receiver<Self::Interface>) {
        tokio::spawn(async move {
            while let Some(message) = rx.recv().await {
                // handle the message
            }
        });
    }
}

let addr = MyService.start();

Required Associated Types

type Interface: Interface

The interface of messages this service implements.

The interface can be a single message type or an enumeration of all the messages that can be handled by this service.

Required Methods

fn spawn_handler(self, rx: Receiver<Self::Interface>)

Spawns a task to handle service messages.

Receives an inbound channel for all messages sent through the service’s Addr. Note that this function is synchronous, so that this needs to spawn a task internally.

Provided Methods

fn start(self) -> Addr<Self::Interface>

Starts the service in the current runtime and returns an address for it.