relay_system

Trait ServiceSpawn

pub trait ServiceSpawn {
    // Required method
    fn start_obj(&self, service: ServiceObj);
}

The ServiceSpawn trait allows for starting a Service on an executor that will run them to completion.

Often you want to spawn your service directly via ServiceSpawnExt.

Required Methods

fn start_obj(&self, service: ServiceObj)

Starts the service on the executor.